+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#5](https://github.com/x1B/nbe/issues/5): Allow to select vertices and remove them using DELETE key.
 - [#2](https://github.com/x1B/nbe/issues/2): Added undo/redo handling.
 - [#4](https://github.com/x1B/nbe/issues/4): Allow to remove selected edge using DELETE key.
 - [#1](https://github.com/x1B/nbe/issues/1): Replaced embedded dependencies with bower.